chore(project): unnest and flatten configs to reduce complexity
