+++ .idea/dictionaries/fuxing.xml
@@ -5,6 +5,7 @@
       <w>fuxing</w>
       <w>fuxingloh</w>
       <w>jsonrpc</w>
+      <w>mainnet</w>
       <w>testcontainers</w>
     </words>
   </dictionary>
