+++ package.json
@@ -13,22 +13,19 @@
     "*": "prettier --write --ignore-unknown"
   },
   "devDependencies": {
-    "@jest/globals": "^29.7.0",
-    "@swc/core": "^1.5.29",
-    "@swc/jest": "^0.2.36",
     "@types/node": "^20.14.2",
     "@workspace/eslint-config": "workspace:*",
-    "@workspace/jest-preset": "workspace:*",
-    "@workspace/prettier-config": "workspace:*",
+    "@workspace/jest": "workspace:*",
     "@workspace/tsconfig": "workspace:*",
     "eslint": "^8.57.0",
     "husky": "^9.0.11",
     "jest": "29.7.0",
     "lint-staged": "^15.2.7",
     "prettier": "^3.3.2",
+    "prettier-plugin-packagejson": "^2.5.0",
+    "prettier-plugin-tailwindcss": "^0.6.4",
     "turbo": "^2.0.4",
-    "typescript": "5.4.5",
-    "wait-for-expect": "^3.0.2"
+    "typescript": "5.4.5"
   },
   "packageManager": "pnpm@8.15.8",
   "engines": {
