+++ packages/chainfile-agent/package.json
@@ -24,7 +24,7 @@
     ]
   },
   "jest": {
-    "preset": "@workspace/jest-preset"
+    "preset": "@workspace/jest"
   },
   "dependencies": {
     "@chainfile/schema": "workspace:^",
