+++ packages/chainfile-schema/package.json
@@ -33,7 +33,7 @@
     ]
   },
   "jest": {
-    "preset": "@workspace/jest-preset"
+    "preset": "@workspace/jest"
   },
   "dependencies": {
     "ajv": "^8.16.0",
