+++ packages/chainfile-testcontainers/tests/bitcoin-regtest.test.ts
@@ -1,4 +1,4 @@
-import { afterAll, beforeAll, describe, expect, it } from '@jest/globals';
+import { afterAll, beforeAll, describe, expect, it } from '@workspace/jest/globals';
 
 import { ChainfileContainer, ChainfileTestcontainers } from '../src';
 import regtest from './bitcoin-regtest.json';
