+++ prettier.config.js
@@ -1,1 +1,12 @@
-module.exports = require('@workspace/prettier-config');
+/** @type {import('prettier').Config} */
+const config = {
+  printWidth: 120,
+  trailingComma: 'all',
+  singleQuote: true,
+  endOfLine: 'lf',
+  tabWidth: 2,
+  useTabs: false,
+  plugins: [import('prettier-plugin-tailwindcss'), import('prettier-plugin-packagejson')],
+};
+
+module.exports = config;
