+++ workspace/eslint-config-next/package.json
@@ -2,7 +2,7 @@
   "name": "@workspace/eslint-config-next",
   "private": true,
   "main": "./index.js",
-  "dependencies": {
+  "devDependencies": {
     "@next/eslint-plugin-next": "14.2.4",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.29.1",
