+++ workspace/eslint-config/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@workspace/eslint-config",
   "private": true,
-  "dependencies": {
+  "devDependencies": {
     "@eslint/js": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.29.1",
