--- workspace/jest-preset/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@workspace/jest-preset",
-  "private": true
-}
