+++ workspace/jest/globals.d.ts
@@ -1,0 +1,1 @@
+export * from '@jest/globals';
