+++ workspace/jest/jest-preset.js
@@ -5,5 +5,5 @@ module.exports = {
   transform: {
     '^.+\\.(t|j)sx?$': '@swc/jest',
   },
-  testTimeout: 120000,
+  testTimeout: 60000,
 };
