+++ workspace/jest/wait-for.d.ts
@@ -1,0 +1,2 @@
+import waitFor from 'wait-for-expect';
+export default waitFor;
