+++ workspace/jest/wait-for.js
@@ -1,0 +1,2 @@
+import waitFor from 'wait-for-expect';
+export default waitFor;
