--- workspace/prettier-config/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "@workspace/prettier-config",
-  "private": true,
-  "dependencies": {
-    "prettier-plugin-packagejson": "^2.5.0",
-    "prettier-plugin-tailwindcss": "^0.6.4"
-  }
-}
